Fundamental Kotlin - Generic functions.
